A little refinement to to Cache Class.
Check for the existence of Cache->collate_files()'s $files_modified var instead of seeing if it has been declared (Resolves PHP5 warning). Also resolve an issue that would not set the Cache->hash var before trying to write it to the cache file, resulting in no hash stored in the file (and a PHP5 warning on next access).