test templates for robots.txt as well as paired content.
The major change is that the content files, while still retaining their normal formatting, allow for the inclusion of unformatted text and style variables. In the case of the css file, this allows you to create a base template with all your resets and base scaffolding that you plugin your own style to, basically like a hardwired @include()

I see these as sort of complementary to the inclusion of the sitemap folder/template in the default stacey install.