added captioning support, allowing descriptions to be added on a per-asset basis
when creating asset collections, stacey will look for a plain text file named
"captions.stacey". this file contains key/value pairs in the same format as
other stacey content files. the key should be the name of the file you wish
to save a caption for, and with the value being either plain text or markdown.